+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `StableVec::grow()`
 - `StableVec::clear()`
 - `StableVec::from_vec()`
+- `StableVec::extend_from_slice()`
 - `Debug` implementations for `Iter`, `IterMut` and `Keys`
 
 ### Changed